fix!: 64-bit chunk-start offsets - data loss for indices >= 2^32 (4.0.0)
sm_add() of any index >= 2^32 silently succeeded but the bit was lost:
sm_contains() returned false afterward.  Root cause: the chunk-start
offset type was 32-bit (typedef uint32_t __sm_idx_t), so the on-disk
chunk start (and the cursor's cached start) truncated at 2^32 while the
public API is fully 64-bit (SM_IDX_MAX == UINT64_MAX).

  * __sm_idx_t is now uint64_t -- the single point of control for the
    on-disk chunk-start width.  It remains internal to sm.c; the public
    API in sm.h is uint64_t for every bit location.  Per-chunk overhead
    grows 4 -> 8 bytes; the serialized wire version bumps 1 -> 2.
  * sm_scan's callback array widens uint32_t[] -> uint64_t[]: it
    delivers absolute bit positions, which also truncated at 2^32.
    This is a public API change (callbacks take uint64_t[] now).
  * RLE descriptor mask literals get explicit ULL suffixes (they were
    signed, risking UB under shifts); several value-changing integer
    conversions are made explicit.

Breaking on-disk format + scan-callback change -> major bump to 4.0.0.
Verified: the >=2^32 reproducer round-trips; full suite green across
seeds under MALLOC_PERTURB_; clean under ASan+UBSan+diagnostic and
valgrind; -Wall -Wextra -Wpedantic clean on gcc and clang.

+++ tests/test_coverage.c
@@ -893,7 +893,7 @@ CASE(test_select_unset_in_partial_rle)
 
 static size_t g_scan_count = 0;
 static uint32_t g_scan_first = 0;
-static void scan_cb(uint32_t vec[], size_t n, void *aux)
+static void scan_cb(uint64_t vec[], size_t n, void *aux)
 {
     (void)aux;
     if (g_scan_count == 0 && n > 0) g_scan_first = vec[0];
